# project-notes/api-spec.yaml
openapi: 3.1.0
info:
  title: Project Allocation Management System (PAMS) API
  version: 1.12.0
  description: |
    REST API for the Project Allocation Management System.
    All endpoints require a Keycloak-issued OIDC Bearer token unless stated otherwise.
    Obtain tokens via the Keycloak Authorization Code flow at the configured realm (`/realms/pams`).
    Identity resolution uses the `empCode` JWT claim for Keycloak-to-PAMS identity bridging (sub claim verified against employees table with empCode fallback).
    Dates use ISO 8601 `YYYY-MM-DD` format. Timestamps use UTC (`YYYY-MM-DDTHH:MM:SSZ`).

    **Pagination (NFR-18/19):** All list endpoints support `page` (default 1) and `limit` (default 10, max 100).
    **Single-endpoint strategy (NFR-20):** Each resource has one list endpoint that also supports search and filter via query parameters.
    **Role-based response shaping (NFR-21):** Responses are filtered server-side based on the authenticated user's role.
    **Consolidated REST design:** Deactivation is handled via PUT (state change), allocation stop via PATCH, allocation remove via DELETE (soft-delete).
    **Billable flag & project status:** Projects carry a `billable` boolean (defaults to true for Client accounts) and a lifecycle `status` (Upcoming → Active → Completed). Both HR and PM (scoped to own projects) can modify status.
...
  contact:
    name: PAMS Engineering
  license:
    name: Internal Use Only
servers:
  - url: https://api.pams.internal/api/v1
...
security:
  - BearerAuth: []
...
paths:
...
  /accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      operationId: accounts_list
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          schema:
            type: string
          description: Search by account code or name (partial match, case-insensitive).
        - name: isActive
          in: query
          schema:
            type: boolean
          description: Filter by active status. Omit to return all.
        - name: accountType
          in: query
          schema:
            $ref: "#/components/schemas/AccountType"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Paginated list of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponse_AccountSummary"
        "400":
          $ref: "#/components/responses/InvalidSortError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Accounts]
      summary: Create a new account (HR only)
      operationId: accounts_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
            example:
              accountCode: ACC-ACME
              accountName: Acme Corporation
              accountType: Client
      responses:
        "201":
          description: Account created
          headers:
            Location:
              schema:
                type: string
              example: /api/v1/accounts/ACC-ACME
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Account code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://pams.internal/errors/ERR_ACCOUNT_CODE_EXISTS"
                title: Conflict
                status: 409
                detail: "Account code 'ACC-ACME' is already in use."
  /accounts/{accountCode}:
    parameters:
      - $ref: "#/components/parameters/AccountCode"

    get:
      tags: [Accounts]
      summary: Get account by code
      operationId: accounts_getByCode
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Accounts]
      summary: Update account (HR only). Account code is immutable. Set isActive=false to deactivate.
      operationId: accounts_update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: "Account has active projects and cannot be deactivated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://pams.internal/errors/ERR_ACCOUNT_HAS_ACTIVE_PROJECTS"
                title: Unprocessable Entity
                status: 422
                detail: "Account has 3 active project(s). Deactivate or complete them before deactivating the account."
  /accounts/export:
    post:
      tags: [Accounts]
      summary: Export accounts as PDF or Excel (HR, PM)
      description: |
        Returns all matching accounts (no pagination) as a downloadable PDF or Excel file.
        Accepts the same filter and sort query parameters as the list endpoint.
        An optional JSON body may specify a column name map to control which columns appear
        in the export and their display labels.
      operationId: accounts_export
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by account code or name.
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: accountType
          in: query
          schema:
            $ref: "#/components/schemas/AccountType"
        - $ref: "#/components/parameters/SortParam"
        - $ref: "#/components/parameters/ExportExtParam"
      requestBody:
        required: false
        description: |
          Optional column mapping. Keys are response DTO field names (case-insensitive),
          values are display labels for the export header. Only matched columns are included.
          If omitted or empty, all default columns are exported.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportColumnMap"
            example:
              accountCode: "Account Code"
              accountName: "Account Name"
              accountType: "Type"
              isActive: "Active?"
      responses:
        "200":
          description: File download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              example: 'attachment; filename="accounts-2026-03-05.pdf"'
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── PROJECTS ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC Bearer token issued by Keycloak (RS256 signed).
        Obtain via Keycloak Authorization Code flow: `{Authority}/protocol/openid-connect/auth`
        or Client Credentials flow for machine-to-machine.
        Default realm: `https://keycloak.internal/realms/pams`.

  # ── PARAMETERS ─────────────────────────────────────────────────────────────
  parameters:
    AccountCode:
      name: accountCode
      in: path
      required: true
      schema:
        type: string
        maxLength: 50
      example: ACC-ACME
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: "Number of items per page (default: 10, max: 100). Per NFR-18."
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: |
        Sort field. Use `fieldName` for ascending or `-fieldName` for descending.
        Each endpoint has a default sort if omitted. Only one sort field per request.
        Invalid field names return 400 with a ProblemDetails response containing
        `field` (the invalid field) and `allowedFields` (list of valid field names).
      example: "-fromDate"
    ExportExtParam:
      name: ext
      in: query
      required: true
      schema:
        type: string
        enum: [pdf, xls]
      description: "Export format: `pdf` for PDF or `xls` for Excel (.xlsx)."

  # ── RESPONSES ──────────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_UNAUTHORIZED"
            title: Unauthorized
            status: 401
            detail: Authentication token is missing or invalid.
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_FORBIDDEN"
            title: Forbidden
            status: 403
            detail: You do not have permission to perform this action.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_NOT_FOUND"
            title: Not Found
            status: 404
            detail: The requested resource was not found.
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_VALIDATION"
            title: Validation Failed
            status: 400
            errors:
              Percentage:
                - Percentage must be at least 25%.
                - Percentage must be a multiple of 5%.
...
    InvalidSortError:
      description: "Invalid sort field. Returns the field name and list of allowed sort fields."
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/InvalidSortProblem"
          example:
            type: "https://pams.internal/errors/ERR_INVALID_SORT"
            title: "Invalid Sort Parameter"
            status: 400
            detail: "Invalid sort field 'xyz'. Allowed fields: accountCode, accountName, accountType, isActive"
            instance: "/api/v1/accounts"
            field: "xyz"
            allowedFields: ["accountCode", "accountName", "accountType", "isActive"]

  # ── SCHEMAS ────────────────────────────────────────────────────────────────
  schemas:
...
    AccountType:
      type: string
      enum: [Client, Internal, Bench]
      description: Type of account.
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        traceId:
          type: string
    ValidationProblemDetails:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
              description: Field-level validation errors keyed by property name.
...
    InvalidSortProblem:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          properties:
            field:
              type: string
              description: The invalid sort field that was provided.
            allowedFields:
              type: array
              items:
                type: string
              description: List of valid sort field names for this endpoint.
          example:
            type: "https://pams.internal/errors/ERR_INVALID_SORT"
            title: "Invalid Sort Parameter"
            status: 400
            detail: "Invalid sort field 'xyz'. Allowed fields: accountCode, accountName, accountType, isActive, totalActiveProjects, totalInactiveProjects"
            field: "xyz"
            allowedFields: ["accountCode", "accountName", "accountType", "isActive", "totalActiveProjects", "totalInactiveProjects"]
    ExportColumnMap:
      type: object
      nullable: true
      additionalProperties:
        type: string
      description: |
        Column name mapping for export. Keys are response DTO field names (case-insensitive match),
        values are the display labels to use in the export header. Only matched columns are included.
        If null or empty, all default columns are exported.
      example:
        empCode: "Employee Code"
        fullName: "Name"

    # ── Pagination ─────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      description: "Standard pagination metadata per NFR-18."
      properties:
        page:
          type: integer
          examples: [1]
        limit:
          type: integer
          examples: [10]
        totalRecords:
          type: integer
          examples: [125]
        totalPages:
          type: integer
          examples: [13]

    # ── Auth ───────────────────────────────────────────────────────────────
    # Login/logout handled by Keycloak directly. No LoginRequest/LoginResponse schemas needed.
    # Authenticated user profile is returned via GET /employees/me (EmployeeDetailResponse).
    # CurrentUserResponse is kept for lightweight "who am I" contexts where full detail is not needed.
...
    AccountSummary:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        accountCode:
          type: string
        accountName:
          type: string
        accountType:
          $ref: "#/components/schemas/AccountType"
        isActive:
          type: boolean
          description: "Whether the account is active. Inactive accounts cannot have new projects created under them."
        totalActiveProjects:
          type: integer
          description: "Computed: count of active projects under this account."
          examples: [3]
        totalInactiveProjects:
          type: integer
          description: "Computed: count of inactive/completed projects under this account."
          examples: [1]
    AccountResponse:
      allOf:
        - $ref: "#/components/schemas/AccountSummary"
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CreateAccountRequest:
      type: object
      required: [accountCode, accountName, accountType]
      properties:
        accountCode:
          type: string
          maxLength: 50
          examples: [ACC-ACME]
        accountName:
          type: string
          maxLength: 150
          examples: [Acme Corporation]
        accountType:
          $ref: "#/components/schemas/AccountType"
    UpdateAccountRequest:
      type: object
      required: [accountName, accountType]
      properties:
        accountName:
          type: string
          maxLength: 150
        accountType:
          $ref: "#/components/schemas/AccountType"
        isActive:
          type: boolean
          description: "Set to false to deactivate. Server returns 422 if account has active projects."
    PagedResponse_AccountSummary:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AccountSummary"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"

    # ── Project ────────────────────────────────────────────────────────────
...